+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Tour: Fixed scroll behaviour when active element is inside modal window
+- TimeRange: Fixed behaviour on Tab click
+- DateRange: Fixed time for date values
 
 
 ## [0.37.0] - 2020-12-21